+++ services/web/server/src/simcore_service_webserver/catalog/_models.py
@@ -3,7 +3,9 @@ from collections.abc import Callable
 from dataclasses import dataclass
 from typing import Any, Final
 
-from aiocache import cached  # type: ignore[import-untyped]
+from aiocache import cached
+from aiohttp import web
+from aiohttp.web import Request
 from models_library.api_schemas_webserver.catalog import (
     ServiceInputGet,
     ServiceInputKey,
@@ -11,7 +13,12 @@ from models_library.api_schemas_webserver.catalog import (
     ServiceOutputKey,
 )
 from models_library.services import BaseServiceIOModel
+from models_library.users import UserID
 from pint import PintError, Quantity, UnitRegistry
+from pydantic import BaseModel, ConfigDict  # type: ignore[import-untyped]
+from servicelib.aiohttp.requests_validation import handle_validation_as_http_error
+
+from ..constants import RQ_PRODUCT_KEY, RQT_USERID_KEY
 
 _logger = logging.getLogger(__name__)
 
@@ -131,3 +138,26 @@ class ServiceOutputGetFactory:
             )
 
         return port
+
+
+class CatalogRequestContext(BaseModel):
+    app: web.Application
+    user_id: UserID
+    product_name: str
+    unit_registry: UnitRegistry
+    model_config = ConfigDict(arbitrary_types_allowed=True)
+
+    @classmethod
+    def create(cls, request: Request) -> "CatalogRequestContext":
+        with handle_validation_as_http_error(
+            error_msg_template="Invalid request",
+            resource_name=request.rel_url.path,
+            use_error_v1=True,
+        ):
+            assert request.app  # nosec
+            return cls(
+                app=request.app,
+                user_id=request[RQT_USERID_KEY],
+                product_name=request[RQ_PRODUCT_KEY],
+                unit_registry=request.app[UnitRegistry.__name__],
+            )
